# backend/src/main/java/com/hourglass/jobclocksync/dto/UserRequest.java
package com.hourglass.jobclocksync.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

@Data
public class UserRequest {
    @NotBlank(message = "Email is required")
    @Email(message = "Invalid email format")
    private String email;
    
    @NotBlank(message = "Password is required")
    private String password;
    
    @NotBlank(message = "Name is required")
    private String name;
    
    @NotNull(message = "Role is required")
    private String role; // VENDOR, COMPANY, ADMIN
    
    private Boolean active = true;
}