# openapi.yml
openapi: 3.0.3
info:
  title: Job Clock Sync API
  version: 1.0.0
  description: |
    REST API for Job Clock Sync - A comprehensive vendor and company management system 
    for job postings, work orders, timesheets, and invoices.
    
    ## Authentication
    All endpoints except `/auth/login` require JWT Bearer token authentication.
    
    ## Roles
    - **VENDOR**: Service providers who can apply for jobs, submit timesheets and invoices
    - **COMPANY**: Organizations who can post jobs, create work orders, and approve submissions
    - **ADMIN**: Administrators who can manage users, roles, and permissions
    
  contact:
    name: Hourglass Team
    email: support@hourglass.com
    url: https://hourglass.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8082/api
    description: Local Development Server
  - url: https://api.hourglass.com
    description: Production Server
...
security:
  - bearerAuth: []
paths:
...
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get details of the currently authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: Successfully retrieved user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ============================================================================
  # User Management APIs (Admin only)
  # ============================================================================
  /users:
    get:
      tags:
        - User Management
      summary: Get all users
      description: Get all users (Admin only)
      operationId: getAllUsers
      responses:
        '200':
          description: Successfully retrieved users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    post:
      tags:
        - User Management
      summary: Create new user
      description: Create a new user (Admin only)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: User already exists or validation error
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}:
    get:
      tags:
        - User Management
      summary: Get user by ID
      description: Get user details by ID (Admin only)
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - User Management
      summary: Update user
      description: Update user details (Admin only)
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - User Management
      summary: Delete user
      description: Delete a user (Admin only)
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/role/{role}:
    get:
      tags:
        - User Management
      summary: Get users by role
      description: Get all users with a specific role (Admin only)
      operationId: getUsersByRole
      parameters:
        - name: role
          in: path
          required: true
          description: User role
          schema:
            $ref: '#/components/schemas/UserRole'
      responses:
        '200':
          description: Successfully retrieved users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid role
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}/toggle-status:
    post:
      tags:
        - User Management
      summary: Toggle user active status
      description: Enable or disable a user account (Admin only)
      operationId: toggleUserStatus
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
      responses:
        '200':
          description: User status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================================
  # Dashboard APIs
  # ============================================================================
...
  /timesheets:
    post:
      tags:
        - Timesheets
      summary: Create timesheet
      description: |
        Create a new timesheet.
        - VENDOR: Creates timesheet for themselves
        - COMPANY: Can create timesheet on behalf of vendor (requires vendorId in request)
      operationId: createTimesheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimesheetRequest'
      responses:
        '200':
          description: Timesheet created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    get:
      tags:
        - Timesheets
      summary: Get timesheets
      description: |
        Get timesheets based on user role:
        - VENDOR: Returns timesheets created by vendor
        - COMPANY: Returns timesheets for company's work orders
      operationId: getTimesheets
      responses:
        '200':
          description: Successfully retrieved timesheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TimesheetResponse'
  /timesheets/{id}/submit:
    post:
      tags:
        - Timesheets
      summary: Submit timesheet
      description: Submit timesheet for approval (Vendor only)
      operationId: submitTimesheet
      parameters:
        - name: id
          in: path
          required: true
          description: Timesheet ID
          schema:
            type: string
      responses:
        '200':
          description: Timesheet submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /timesheets/{id}/approve:
    post:
      tags:
        - Timesheets
      summary: Approve timesheet
      description: Approve a submitted timesheet (Company only)
      operationId: approveTimesheet
      parameters:
        - name: id
          in: path
          required: true
          description: Timesheet ID
          schema:
            type: string
      responses:
        '200':
          description: Timesheet approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /timesheets/{id}/reject:
    post:
      tags:
        - Timesheets
      summary: Reject timesheet
      description: Reject a submitted timesheet (Company only)
      operationId: rejectTimesheet
      parameters:
        - name: id
          in: path
          required: true
          description: Timesheet ID
          schema:
            type: string
      responses:
        '200':
          description: Timesheet rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimesheetResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ============================================================================
  # Invoice APIs
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer Token Authentication. 
        Login via /auth/login to get a token, then include it in the Authorization header as 'Bearer {token}'
  responses:
    UnauthorizedError:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            timestamp: '2025-11-05T10:30:00Z'
            status: 401
            error: Unauthorized
            message: Invalid or expired token
            path: /api/work-orders
    ForbiddenError:
      description: Forbidden - User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            timestamp: '2025-11-05T10:30:00Z'
            status: 403
            error: Forbidden
            message: Access denied
            path: /api/work-orders
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            timestamp: '2025-11-05T10:30:00Z'
            status: 404
            error: Not Found
            message: Resource not found
            path: /api/work-orders/123
  schemas:
...
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: User ID
          example: 507f1f77bcf86cd799439011
        email:
          type: string
          format: email
          description: User email address
          example: company@example.com
        name:
          type: string
          description: User or company name
          example: Acme Corporation
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
          description: Whether the user account is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: '2025-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2025-11-05T10:30:00Z'
    UserRole:
      type: string
      enum:
        - VENDOR
        - COMPANY
        - ADMIN
      description: |
        User role in the system:
        - VENDOR: Service provider who can apply for jobs and submit timesheets/invoices
        - COMPANY: Organization who can post jobs and approve submissions
        - ADMIN: Administrator who can manage users and permissions

    # ==========================================================================
    # Dashboard Schemas
    # ==========================================================================
...
    UserRequest:
      type: object
      required:
        - email
        - password
        - name
        - role
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: newuser@example.com
        password:
          type: string
          format: password
          description: User password
          example: securePassword123
          minLength: 6
        name:
          type: string
          description: User or company name
          example: John Doe
          minLength: 1
        role:
          type: string
          description: User role (VENDOR, COMPANY, or ADMIN)
          example: VENDOR
          enum:
            - VENDOR
            - COMPANY
            - ADMIN
        active:
          type: boolean
          description: Whether the user account is active
          example: true
          default: true
    UserResponse:
      type: object
      required:
        - id
        - email
        - name
        - role
        - active
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: User ID
          example: 507f1f77bcf86cd799439011
        email:
          type: string
          format: email
          description: User email address
          example: user@example.com
        name:
          type: string
          description: User or company name
          example: John Doe
        role:
          $ref: '#/components/schemas/UserRole'
        active:
          type: boolean
          description: Whether the user account is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
          example: '2025-01-01T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2025-11-05T10:30:00Z'

    # ==========================================================================
    # Timesheet Schemas
    # ==========================================================================
    TimesheetRequest:
      type: object
      required:
        - workOrderId
        - weekStartDate
        - weekEndDate
        - entries
      properties:
        workOrderId:
          type: string
          description: Associated work order ID
          example: 507f1f77bcf86cd799439013
        weekStartDate:
          type: string
          format: date
          description: Week start date (YYYY-MM-DD)
          example: '2025-11-01'
        weekEndDate:
          type: string
          format: date
          description: Week end date (YYYY-MM-DD)
          example: '2025-11-07'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/TimesheetEntryRequest'
          description: List of timesheet entries
          minItems: 1
        notes:
          type: string
          description: Additional notes
          example: Completed initial setup and testing
        vendorId:
          type: string
          description: |
            Optional vendor ID when company creates timesheet on behalf of vendor.
            Required when request is made by COMPANY role.
          example: 507f1f77bcf86cd799439012
    TimesheetEntryRequest:
      type: object
      required:
        - date
        - hours
        - description
        - workOrderId
      properties:
        date:
          type: string
          format: date
          description: Entry date (YYYY-MM-DD)
          example: '2025-11-01'
        hours:
          type: number
          format: double
          description: Hours worked
          example: 8.0
          minimum: 0
        description:
          type: string
          description: Work description
          example: Initial setup and configuration
          minLength: 1
        workOrderId:
          type: string
          description: Associated work order ID
          example: 507f1f77bcf86cd799439013
    TimesheetResponse:
      type: object
      required:
        - id
        - vendorId
        - companyId
        - workOrderId
        - status
        - weekStartDate
        - weekEndDate
        - entries
        - totalHours
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          description: Timesheet ID
          example: 507f1f77bcf86cd799439017
        vendorId:
          type: string
          description: Vendor who submitted the timesheet
          example: 507f1f77bcf86cd799439012
        companyId:
          type: string
          description: Company that will approve the timesheet
          example: 507f1f77bcf86cd799439011
        workOrderId:
          type: string
          description: Associated work order
          example: 507f1f77bcf86cd799439013
        status:
          $ref: '#/components/schemas/TimesheetStatus'
        weekStartDate:
          type: string
          format: date
          description: Week start date
          example: '2025-11-01'
        weekEndDate:
          type: string
          format: date
          description: Week end date
          example: '2025-11-07'
        entries:
          type: array
          items:
            $ref: '#/components/schemas/TimesheetEntryResponse'
          description: List of timesheet entries
        totalHours:
          type: number
          format: double
          description: Total hours across all entries
          example: 40.0
        notes:
          type: string
          nullable: true
          description: Additional notes
          example: Completed initial setup and testing
        submittedDate:
          type: string
          format: date-time
          nullable: true
          description: Submission timestamp
          example: '2025-11-08T09:00:00Z'
        approvedDate:
          type: string
          format: date-time
          nullable: true
          description: Approval timestamp
          example: '2025-11-09T14:30:00Z'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: '2025-11-05T10:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2025-11-08T09:00:00Z'
    TimesheetEntryResponse:
      type: object
      required:
        - date
        - hours
        - description
        - workOrderId
      properties:
        date:
          type: string
          format: date
          description: Entry date
          example: '2025-11-01'
        hours:
          type: number
          format: double
          description: Hours worked
          example: 8.0
        description:
          type: string
          description: Work description
          example: Initial setup and configuration
        workOrderId:
          type: string
          description: Associated work order ID
          example: 507f1f77bcf86cd799439013
    TimesheetStatus:
      type: string
      enum:
        - DRAFT
        - SUBMITTED
        - APPROVED
        - REJECTED
      description: |
        Timesheet status:
        - DRAFT: Timesheet is being drafted
        - SUBMITTED: Timesheet submitted for approval
        - APPROVED: Timesheet approved by company
        - REJECTED: Timesheet rejected by company

    # ==========================================================================
    # Invoice Schemas
    # ==========================================================================
...
    Error:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2025-11-05T10:30:00Z'
        status:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type
          example: Bad Request
        message:
          type: string
          description: Error message
          example: Validation failed
        path:
          type: string
          description: Request path that caused the error
          example: /api/work-orders